Show contextual filters in Search Results Table
Closes #89
